service/requirements.txt
google-cloud-pubsub==1.4.2
flask==1.1.1
cherrypy==18.3.0